requirements.txt
# Core dependencies for beach object detection training
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Computer vision and image processing
opencv-python>=4.8.0
Pillow>=9.0.0

# Data manipulation and analysis
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars and utilities
tqdm>=4.65.0

# Configuration and data serialization
PyYAML>=6.0

# Optional: For advanced hyperparameter tuning and experiments
scikit-learn>=1.3.0

# Development and testing (optional)
pytest>=7.0.0
